requirements.txt
openai>=1.0.0
tqdm
google-generativeai
torch
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
datasets
trl>=0.7.0
